Update controllers and views

+++ app/controllers/books_controller.rb
@@ -3,8 +3,7 @@ class BooksController < ApplicationController
   expose :books, -> { Book.all }
   expose :book
 
-  before_action :set_user, only: %i[edit update destroy]
-  before_action :require_admin, only: %i[edit update destroy]
+  # before_action :set_user, only: %i[edit update destroy]
   before_action :set_reader, only: [:show]
   before_action :set_future_reader, only: [:show]
   # before_action only: %i[show edit update destroy] do
@@ -27,7 +26,12 @@ class BooksController < ApplicationController
     end
   end
 
+  def edit
+    authorize book
+  end
+
   def update
+    authorize book
     if book.update(book_params)
       flash[:notice] = 'Book was successfully updated.'
       redirect_to book
@@ -37,6 +41,7 @@ class BooksController < ApplicationController
   end
 
   def destroy
+    authorize book
     book.destroy
     flash[:notice] = 'Book was successfully deleted.'
     redirect_to books_url
