+++ app/controllers/future_readers_controller.rb
@@ -1,13 +1,14 @@
 # HANDLE USERS ON THE WISHLIST
 class FutureReadersController < ApplicationController
-  expose :books, ->{ Book.all }
+  expose :books, -> { Book.all }
   expose :book
 
   # before_action :set_book, only: %i[create show destroy]
   before_action :set_reader, only: %i[create show destroy]
-  before_action :set_user, only: %i[create show destroy]
+  # before_action :set_user, only: %i[create show destroy]
 
   def create
+    authorize book, policy_class: ReadablePolicy
     @future_reader = current_user.future_readers.build(future_readers_params)
     # @future_reader.book_id = book.id
     respond_to do |format|
@@ -27,11 +28,13 @@ class FutureReadersController < ApplicationController
     end
   end
 
-  # def show
+  def show
+    authorize book, policy_class: ReadablePolicy
   #   destroy
-  # end
+  end
 
   def destroy
+    authorize book, policy_class: ReadablePolicy
     @future_reader = FutureReader.book_read_by(book, current_user)
     respond_to do |format|
       if !@future_reader.nil? && @future_reader.destroy
